% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_cpp}
\alias{add_cpp}
\title{Add two integers using C++}
\usage{
add_cpp(x, y)
}
\arguments{
\item{x}{First integer.}

\item{y}{Second integer.}
}
\value{
The sum of x and y.
}
\description{
Add two integers using C++
}
