% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_without_para.R
\name{brute_without_para}
\alias{brute_without_para}
\title{Brute without para}
\usage{
brute_without_para(x, W)
}
\arguments{
\item{x}{is a data frame}

\item{W}{is a integer}
}
\value{
will return a list
}
\description{
Brute without para
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <-data.frame(
  w=sample(1:4000, size = n, replace = TRUE),
  v=runif(n = n, 0, 10000)
)
brute_without_para(x = knapsack_objects[1:8,], W = 3500)
}
