% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bruteforce_knapsack}
\alias{bruteforce_knapsack}
\title{Brute Force method(exhaustive search) to solve knapsack problem in C++}
\usage{
bruteforce_knapsack(set_of_weight, set_of_profit, W_max, R_number)
}
\arguments{
\item{set_of_weight}{weights}

\item{set_of_profit}{profits}

\item{W_max}{maximum weight}

\item{R_number}{passed from R}
}
\value{
A list with optimal value and selected elements
}
\description{
Brute Force method(exhaustive search) to solve knapsack problem in C++
}
