% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6function.R
\name{dynamic_programming}
\alias{dynamic_programming}
\title{dynamic_programming}
\usage{
dynamic_programming(x, W, fast = FALSE)
}
\arguments{
\item{x}{,a data.frame. x$w is the weight,x$v is the value}

\item{W}{,a integer. The maximum weight of items in the bag.}

\item{fast, }{a logical variable. When it is set TRUE,then it will call Rcpp to solve.}
}
\value{
list , returns a list containing the maximum value and the index of items in x.
}
\description{
dynamic_programming for lab6
}
