\name{FastDynamicProgramming}
\alias{FastDynamicProgramming}
\title{
Simple function using Rcpp
}
\description{
Simple function using Rcpp
}
\usage{
FastDynamicProgramming(x,W)	
}
\arguments{
\item{x}{,a data.frame. x$w is the weight,x$v is the value}

\item{W}{,a integer. The maximum weight of items in the bag.}
}
\examples{
\dontrun{
FastDynamicProgramming()
}
}
