% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_knapsack_objects.R
\name{generate_knapsack_objects}
\alias{generate_knapsack_objects}
\title{Generate Knapsack Objects}
\arguments{
\item{n}{An integer specifying the number of knapsack objects to generate. Default is 2000.}
}
\value{
A data frame with two columns: \code{w} (weights) and \code{v} (values) of the knapsack items.
}
\description{
This function generates a data frame with random weights and values for knapsack items.
The number of items is specified by the user.
}
\examples{
generate_knapsack_objects(10)
}
