% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\name{get_indexes}
\alias{get_indexes}
\title{get_indexes
 
 get_indexes by using the combinatinons and knapsack data}
\usage{
get_indexes(vector, data)
}
\arguments{
\item{vector}{input multiple_combinations[i, ]}

\item{data}{the knapsack data}
}
\description{
get_indexes
 
 get_indexes by using the combinatinons and knapsack data
}
