% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\name{get_value}
\alias{get_value}
\title{get value
 
 get value by using the combinatinons and knapsack data}
\usage{
get_value(vector, data)
}
\arguments{
\item{vector}{input multiple_combinations[i, ]}

\item{data}{the knapsack data}
}
\description{
get value
 
 get value by using the combinatinons and knapsack data
}
