% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnapsackBruteForce.R
\name{getAllSubsets}
\alias{getAllSubsets}
\title{getAllStatus}
\usage{
getAllSubsets(n)
}
\arguments{
\item{n}{integer}
}
\value{
list of all subsets
}
\description{
Helper function getAllSubsets(n), returns all subsets of a value n
}
