% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{#============Gredy
#' greedy_knapsack
#'
#' Using Greedy algorithm to solve
#' @param x the data set Knapsack_object
#' @param W the max weight allowed
#'
#' @export
#' @importFrom utils tail}
\usage{
greedy_knapsack(x, W)
}
\description{
#============Gredy
#' greedy_knapsack
#'
#' Using Greedy algorithm to solve
#' @param x the data set Knapsack_object
#' @param W the max weight allowed
#'
#' @export
#' @importFrom utils tail
}
