% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_heuristic2.r
\name{greedy_knapsack2}
\alias{greedy_knapsack2}
\title{Implements Greedy approach for knapsack problem}
\source{
https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm
}
\usage{
greedy_knapsack2(x, W)
}
\arguments{
\item{x}{data.frame,with a weight w and a value v}

\item{W}{Knapsack capacity}
}
\value{
Maximum knapsack value and which elements (row in x)
}
\description{
a greedy approximation algorithm which solves the unbounded knapsack problem
}
\details{
This algorithm will not give an exact result,it will reduce the computational complexity considerably
}
