% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack_improved.R
\name{greedy_knapsack_improved}
\alias{greedy_knapsack_improved}
\title{Greedy Heuristic approach for the knapsack problem}
\usage{
greedy_knapsack_improved(x, W)
}
\arguments{
\item{x}{data frame with two columns \code{w} (weight) and \code{v} (value) of items to place in the knapsack}

\item{W}{the maximum weight (numeric) the knapsack can hold}
}
\value{
theoretical maximum \code{$val} (knapsack value) composed of \code{$ids} (which items)
}
\description{
This modified algorithm does not stop when getting an item that cannot get inside the knapsack, so its "improved"
}
