% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_programming}
\alias{greedy_programming}
\alias{greedy_knapsack}
\title{greedy programming}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{dataframe}

\item{W}{number}
}
\description{
This approach is to use the a heuristic or approximation for the problem
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000))
 greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
\references{
https://en.wikipedia.org/wiki/Knapsack problem#Greedy approximation algorithm
}
\author{
Mahnaz , Bita
}
