% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_brute_force.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{Brute force solution to the knapsack problem}
\usage{
knapsack_brute_force(x, W)
}
\value{
The function returns the maximum value and all possible wheight combinations of this solution.
}
\description{
A solution to the knapsack problem that employs a bruteforce apporach to the problem. 
The function generates and test for all possible combinations. This generates a huge
amount of calculations for larger data-sets. Users are not recomended to give the 
function a data material with more than 20 rows.
}
\examples{

set.seed(42)
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 )

brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)

brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)



}
\references{
\link{https://en.wikipedia.org/wiki/Knapsack_problem}
}

