% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{A dynamic programming solution to 0/1 Knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data.frame with two columns named in order \code{"w"} and \code{"v"}
 Every row in \code{x} is an object with weight \code{w} and value \code{v}.}

\item{W}{A numeric scalar which is the limit of the weight the knapsack can carry.}
}
\value{
\code{list} A list with names \code{$value}, telling the maximum value of the knapsack
 and \code{$elements} which indicates which row objects in data.frame \code{x} was put
 in the knapsack.
}
\description{
This function solves the 0/1 Knapsack problem using dynamic programming
}
\details{
The Knapsack problem is a combinatorial optimization problem
where one tries to fill a limited weight knapsack with objects
of as high a total value as possible. Every object has a positive
weight and value associated with itself.
}
\examples{
knapsack_dynamic(data.frame(w=c(20,30,40,50),v=c(2,2,1,2)),W =20)
}
\references{
\url{http://en.wikipedia.org/wiki/Knapsack_problem}
}

