% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knapsack_gener_matr.R
\name{knapsack_gener_matr}
\alias{knapsack_gener_matr}
\title{Generates values for knapsack problem}
\usage{
knapsack_gener_matr()
}
\value{
\code{knapsack_objects} A dataframe with two columns, w and v
containing weights and values for use in the knapsack problem.
}
\description{
This function is the knapsack value generator in the Lab instructions
for Lab 6 in 732A50. It is the same as the \code{knapsack_generator} function but this one uses
matrix instead of data frame.
}

