% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knapsack_generator.R
\name{knapsack_generator}
\alias{knapsack_generator}
\title{Generates values for knapsack problem}
\usage{
knapsack_generator()
}
\value{
\code{knapsack_objects} A dataframe with two columns, w and v
containing weights and values for use in the knapsack problem.
}
\description{
This function is the knapsack value generator in the Lab instructions
for Lab 6 in 732A50.
}

