% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6-package.R
\docType{package}
\name{Lab6-package}
\alias{Lab6}
\alias{Lab6-package}
\title{Lab6: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Jonatanprepuk/Lab6}
  \item Report bugs at \url{https://github.com/Jonatanprepuk/Lab6/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
