% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\docType{package}
\name{Lab6}
\alias{Lab6}
\title{Lab6:: A package for implementing solutions for the knapsack problem}
\description{
With three different methods execution times calculated and compared. Also we profile our code with parallel method and discuss the results about the gained time.
}
\section{Brute Force}{

Brute force method is used to see the execution time of the problem.
}

\section{Dynamic knapsack function}{

Dynamic method is used to see the execution time of the problem.
}

\section{Greedy knapsack}{

Greedy method is used to see the execution time of the problem.
}

