% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized_brute_force_knapsack.R
\name{optimized_brute_force_knapsack}
\alias{optimized_brute_force_knapsack}
\title{Optimized Brute Force Knapsack Solver}
\usage{
optimized_brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with two variables, the value of the item and the weight of the item.}

\item{W}{A numeric number that sets the total weight capacity of the knapsack and must be greater than 0}
}
\value{
A list of the selected items total value and a vector of the selected items observation number
}
\description{
This function solves the knapsack problem by a Brute force search and look
through every combination of items to find the max value without overstep the
weight limit. It is optimized using vectorisation to remove the outer for loop.
}
\examples{
# Example with a dataset
RNGversion(min(as.character(getRversion()),"3.5.3"))
##old sampler used for backward compatibility
# suppressWarnings() can be used so that the above warning is not displayed
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-data.frame(
 w=sample(1:4000, size = n, replace = TRUE),
v=runif(n = n, 0, 10000)
)

optimized_brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)


}
