% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profiling.R
\name{profiling}
\alias{profiling}
\title{This function measures performance of code}
\usage{
profiling()
}
\description{
Some time trials of our code
}
\details{
We use Rprof to find out performance of certain bits of code in this package.
We also found out two nested for loops are a little faster than two nested
while loops.
}

