% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomData.R
\name{RandomData}
\alias{RandomData}
\alias{randomData}
\title{RandomData}
\usage{
randomData(data_length = NA)
}
\arguments{
\item{data_length}{A numeric variable, default NA.}
}
\value{
A data frame who contains w for weights and v for values.
}
\description{
This file aims to generate input data for knapsack.
}
\examples{
knapsack_objects <- randomData(data_length = 8)
}
