% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicProgramming.R
\name{updateStatus}
\alias{updateStatus}
\title{updateStatus}
\usage{
updateStatus(not_load, load, value)
}
\arguments{
\item{not_load}{A numeric variable.}

\item{load}{A numeric variable.}

\item{value}{A numeric variable.}
}
\value{
A numeric element.
}
\description{
This function aims to improve the performance in R implemented by C++.
}
