% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_brute.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Algorithm}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{is a data frame containing weights and values of the elements.}

\item{W}{is the highest weight capacity.}
}
\value{
output is the maximum value the knapsack can hold for the given weight capacity
}
\description{
Knapsack problem is solved using brute force algorithm.
}
\examples{
brute_force(x = knapsack_objects[1:5,], W = 150)

}
\references{
\href{https://en.wikipedia.org/wiki/Knapsack_problem}{Dynamic Algorithm}
}
