% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruteForce.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{It represents the input data frame with 2 columns w (weights) and v (values).}

\item{W}{It represents the knapsack size.}

\item{parallel}{It determines whether parallel computation is used and it is logical.}
}
\value{
list It contains the maximum value and the elements selected to maximize the knapsack value.
}
\description{
This function implements the brute force knapsack.
}
