% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute-force search is guaranteed to give a correct answer in all situations for the knapsack problem.}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x, }{data.frame with with two variables weight(w) and value(v)}

\item{W, }{the size of knapsack}
}
\value{
the maximum knapsack value and which elements
}
\description{
This function calculates all possible alternatives and return the maximum value.
}
\details{
This function calculates in the easiest way to enumerate all different combinations.
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}

