% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force algorithm goes through all possible alternatives and returns the 
maximum value found.}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data frame with two variables, x = value and w = weight}

\item{W}{knapsack size}
}
\value{
the maximum knapsack value and related elements
}
\description{
This approach is of complexity O(2^n).
}
\seealso{
\url{https://en.wikipedia.org/wiki/Knapsack}
}
