% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruteknapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{This is a description of brute_force_knapsack function}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{a data frame has two columns and all the elements are positive numbers}

\item{W}{a single number denotes the size of the knapsack}
}
\value{
a list contains the biggest value and the items
}
\description{
This function creates a brute forced algorithm to solve the knapsack problem.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n , replace = TRUE),
v = runif(n = n, 0 , 10000))
l <- brute_force_knapsack(x = knapsack_objects[1:8, ], W = 3500)

}
