% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Knapsack Solver}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data.frame with two variables v and w}

\item{W}{the knapsack size}
}
\value{
A list with:
\describe{
\item{value}{The maximum total value found.}
\item{elements}{The indices of the selected items.}
}
}
\description{
Brute Force Knapsack Solver
}
\examples{
# Example with 8 items
n <- 8
knapsack_objects <- data.frame(
  w = sample(1:4000, size = n, replace = TRUE),
  v = runif(n = n, 0, 10000)
)
brute_force_knapsack(x = knapsack_objects[1:8, ], W = 3500)

}
