% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force knapsack

The brute force knapsack algorithm}
\arguments{
\item{x}{data frame with value and weights}

\item{W}{the capacity of the knapsack}
}
\value{
total value and elements
}
\description{
calculates the knapsack problem with a brute force algorithm.
}
\references{
https://en.wikipedia.org/wiki/Knapsack_problem
}
