% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{is a data frame}

\item{W}{given max weight}

\item{parallel}{logical value}
}
\value{
a list
}
\description{
Implement knapsack by appling brute force which is actually we obtained
by making subset and then witht help help of parallel we are saving time.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <-data.frame(
  w=sample(1:4000, size = n, replace = TRUE),
  v=runif(n = n, 0, 10000)
)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500,parallel=FALSE)
}
