% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_brute_force.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force method(exhaustive search) to solve knapsack problem}
\usage{
brute_force_knapsack(x, W, methhod = "Not_so_fast")
}
\arguments{
\item{x}{A data frame with two elements, weights(w) and values(v)}

\item{W}{Knapsack capacity}

\item{methhod}{fast Not_so_fast}
}
\value{
A list with optimal value and selected elements
}
\description{
Brute Force method(exhaustive search) to solve knapsack problem
}
