% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Algorithm for Knapsack Problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} with values and weights
for all items. Each row represents one item and must contain
the following columns with positive values:
\describe{
\item{\code{w}}{A \code{numeric} representing the weight.}
\item{\code{v}}{A \code{numeric} representing the value.}
}}

\item{W}{A positive \code{\link[base]{numeric}} giving the capacity
of the knapsack.}
}
\value{
A \code{\link[base]{list}} with elements:
\describe{
\item{\code{value}}{Total value of the optimal knapsack.}
\item{\code{elements}}{Row indices in \code{x} of the items
included in the knapsack.}
}
}
\description{
Brute Force Algorithm for Knapsack Problem
}
\examples{
x <- data.frame(w = c(2, 3, 4), v = c(3, 4, 5))
knapsack_brute_force(x, W = 5)

}
