% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_brute_force.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Search}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{Data Frame that contain v(value) and w(weight)}

\item{W}{Vector eith the knapsack size}
}
\value{
a list with the value and elements of the knapsack
}
\description{
use the brute force search to solves the problem
}
\examples{
set.seed(42) 
n <- 2000 
knapsack_objects <- data.frame( w=sample(1:4000, size = n, replace = TRUE), v=runif(n = n, 0, 10000))
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)


}
