% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute_force_knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A data frame that represents the wights and the values of each item.}

\item{W}{A number that represents the total weight .}

\item{parallel}{A logical value that indicate if you want to use parallel .}
}
\value{
A list with the optimum value and the number of each item.
}
\description{
Brute_force_knapsack
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
