% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack_parallel}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{dataframe}

\item{W}{number}

\item{parallel}{logical}
}
\description{
This approach is to use the a heuristic or approximation for the problem
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000))
 brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
}
\author{
Mahnaz , Bita
}
