% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force Solution for Knapsack Problem}
\usage{
brute_force_knapsack(x,W,parallel)
}
\arguments{
\item{x}{is a data frame with two columns w(weight) and v(value)}

\item{W}{is maximum weight(capacity) of knapsack}

\item{parallel}{is to tell R whether or not to use parallel computation}
}
\value{
\code{list} List of object containing \code{value} giving maximum value of Knapsack out of dataframe and \code{elements} giving weight of 
selected elements from data frame x
}
\description{
The knapsack problem is a combinatorial optimization problem.
A dataframe is given having two parameters, weight and value. 
Each value has its own weight and we have to pack items with maximum value and within weight capacity.
This function will run in O<2^n>.
}
\examples{
  RNGversion(min(as.character(getRversion()),"3.6.1"))
  set.seed(42,kind="Mersenne-Twister",normal.kind = "Inversion")
  n <- 2000
  knapsack_objects <-data.frame(w=sample(1:4000, size = n, replace = TRUE),
                             v=runif(n = n, 0, 10000))
  l<-brute_force_knapsack(knapsack_objects[1:16,] , 3500,FALSE)
  
}
\seealso{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
