% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force algorithm for the knapsack problem}
\source{
https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm
}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data.frame,with a weight w and a value v}

\item{W}{Knapsack capacity}
}
\value{
Maximum knapsack value and which elements (row in x)
}
\description{
takes a data.frame cx with two variables v and w and returns the maximum knapsack value and which elements (rows in the data.frame)
}
\details{
takes a data.frame cx with two variables v and w and returns the maximum knapsack value and which elements (rows in the data.frame)
}
