% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Algorithm for the Knapsack problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data frame with two columns \code{w}(weight) and \code{v}(value) of items}

\item{W}{maximum weight the knapsack can hold}
}
\value{
theoretical maximum \code{$value} (knapsack value) composed of \code{$elements} (which items)
}
\description{
With this algorithm we can solve the Knapsack problem in the brute force way, this approach is of complexity O(2^n)
}
