% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsakh brute force search}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A data frame containing values and weights of items}

\item{W}{The knapsack's capacity}
}
\value{
A data frame of two elements: maximum value and selected elements.
}
\description{
Knapsakh brute force search
}
