% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{An implementation of the brute force solution to the knapsack problem
Where we list all possible combinations of a vector with size \{0,1\}\*n and then make it into
a matrix and check each row for best possible value. The matrix has the size 2^n rows and n columns.}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data.frame with two vectors of the same size, v with values and w with weights}

\item{W}{an integer as the total weight of the knapsack}

\item{parallel}{boolean deciding if computation should be done in parallel or concurrent}
}
\description{
An implementation of the brute force solution to the knapsack problem
Where we list all possible combinations of a vector with size \{0,1\}\*n and then make it into
a matrix and check each row for best possible value. The matrix has the size 2^n rows and n columns.
}
