% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A Data frame (knapsack)}

\item{W}{An integer(Capacity)}
}
\description{
Helps you speed up the code
}
\details{
Brute Force Search
}
