% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnapsackBruteForce.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{knapsack_brute_force}
\usage{
brute_force_knapsack(x = NA, W = NA)
}
\arguments{
\item{x}{data.frame x with two variables v and w}

\item{W}{The variable is the knapsack size}
}
\value{
returns the maximum knapsack value and which elements
}
\description{
This class uses the Kolada API
}
\examples{
knapsack_objects <- randomData(data_length = 8)
brute_force_knapsack(x = knapsack_objects, W = 3500)
}
