% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnapsackBruteForce.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x = NA, W = NA)
}
\arguments{
\item{x}{data.frame x with two variables v and w}

\item{W}{The variable is the knapsack size}
}
\value{
returns the maximum knapsack value and which elements
}
\description{
This Function solves the knapsack poblem by brute force
}
\examples{
output <- brute_force_knapsack(list(v <- c(1,2,3), w <- (4,5,6)), W = 10)
}
