% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6function.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{lab6 assignment}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{,a data.frame. x$w is the weight,x$v is the value}

\item{W}{,a integer. The maximum weight of items in the bag.}
}
\value{
list , returns a list containing the maximum value and the index of items in x.
}
\description{
brute_force_knapsack for lab6
}
