% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsacks}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{the data set Knapsack_object}

\item{W}{the max weight allowed}

\item{parallel}{Using parallel or not, defult is FALSE}
}
\description{
Using brute force to solve knapsack
}
