% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Largest possible value in weighted bag - brute force}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data frame}

\item{W}{positive numeric value}

\item{parallel}{gives opportunity to use parallel programming}
}
\value{
The output will be a numerical value(the largest possible sum) and the corresponding numerical elements that make that sum
}
\description{
This function takes the values and weights provided in a data frame and finds the largest possible sum of values that is less than or equal to the inputed weight by trying all possible combinations.
}
