% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack_parallel.R
\name{brute_force_knapsack_parallel}
\alias{brute_force_knapsack_parallel}
\title{Brute Force Knapsack Solver (with optional parallelization)}
\usage{
brute_force_knapsack_parallel(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data.frame with two variables v and w}

\item{W}{the knapsack size}

\item{parallel}{logical; if TRUE, the search will be parallelized across available cores.}
}
\value{
A list with:
\describe{
\item{value}{The maximum total value found.}
\item{elements}{The indices of the selected items.}
}
}
\description{
Brute Force Knapsack Solver (with optional parallelization)
}
\examples{
n <- 8
knapsack_objects <- data.frame(
  w = sample(1:4000, size = n, replace = TRUE),
  v = runif(n = n, 0, 10000)
)
brute_force_knapsack_parallel(x = knapsack_objects[1:8, ], W = 3500, parallel = TRUE)

}
