% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Dynamic programming}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{data frame with two columns:  v = value and w = weights of each item to put it into the knapsack. It is of complexity O(Wn), so it should scale much better than the brute force alogrithm.}

\item{W}{knapsack size, needs to be a positive value}
}
\value{
the maximum knapsack value and related elements
}
\description{
This algorithm tries to solve the knapsack problem, by an dynamic approach.
}
\examples{
library(Lab6)
dynamic_knapsack(x = knapsack_objects[1:8,], W = 3500)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Knapsack}
}
