% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Dynamic programming}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{data frame with two variables, x = value and  w= weight}

\item{W}{knapsack size}
}
\value{
the maximum knapsack value and related elements
}
\description{
xxxx
}
\seealso{
\url{https://en.wikipedia.org/wiki/Knapsack}
}
