% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{knapsack_dynamic

The dynamic knapsack algorithm}
\arguments{
\item{x}{data frame with value and weights}

\item{W}{the capacity of the knapsack}
}
\value{
total value and elements
}
\description{
calculates the knapsack problem with a dynamic algorithm.
}
\references{
https://en.wikipedia.org/wiki/Knapsack_problem#Dynamic_programming_in-advance_algorithm
}
