% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{Dynamic_programming}
\alias{Dynamic_programming}
\alias{knapsack_dynamic}
\title{Dynamic programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{dataframe}

\item{W}{number}
}
\description{
This approach is of complexity O(2n) since all possible combinations 2n needs to be evaluated
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000))
}
\references{
https://en.wikipedia.org/wiki/Knapsack problem#0.2F1 knapsack problem
}
\author{
Mahnaz , Bita
}
