% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{is a data frame containing weights and values of elements.}

\item{W}{is the highest weight capacity.}
}
\value{
the output is the maximum value the knapsack can holf for the give weight capacity.
}
\description{
Knapsack problem is solved using greedy algorithm.
}
\examples{
greedy_knapsack(x = knapsack_objects[1:100,], W = 250)

}
\references{
\href{https://en.wikipedia.org/wiki/Knapsack_problem}{Dynamic Algorithm}
}
