% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic in knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe with two columns: the values (v) and the weights (w) of each item to put in the knapsack.}

\item{W}{A positive number representing the knapsack size.}
}
\value{
A list of two elements: a positive number with the maximum knapsack \code{value} and a vector of all the \code{elements} in the knapsack size.
}
\description{
Greedy heuristic in knapsack problem
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <-data.frame(w=sample(1:4000, size = n, replace = TRUE),v=runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
\references{
\url{ https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm }
}
