% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack Solver}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with two columns, 'v' for values and 'w' for weights.}

\item{W}{The capacity of the knapsack.}
}
\value{
A list with the maximum value and a vector of selected item indices.
}
\description{
This function solves the knapsack problem using a greedy approach. In this approach, items are sorted by their value-to-weight ratio in descending order and then selected one by one until the knapsack's weight limit is reached.
}
