% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{A greedy solution to the knapsack_problem}
\usage{
greedy_knapsack(x, W)
}
\value{
It returns an aproximate maximum value and what elements this value is based of.
}
\description{
This function implements a greedy aproximation algorithm to the knapsack problem. The function 
takes a max weight W and a data.frame x that contains a column v (value) for  and column w (weight).
}
\examples{

## Generate data

set.seed(42)
n <- 2000 
knapsack_objects <- data.frame( w=sample(1:4000, size = n, replace = TRUE), 
                               v=runif(n = n, 0, 10000) )

greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)


greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)


}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}

