% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Greedy_heuristic.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic algorithm for the knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{is a data.frame containing variables 'w' and 'v', weights and values}

\item{W}{is the total size of the knapsack}
}
\description{
Greedy heuristic algorithm for the knapsack problem
}
\details{
This algorithm is approximative, it can be shown that it will return at least 50 % of the true maximum value. The value of this algorithm lies in the low computational time
}
\examples{
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}
\seealso{
\code{\link{brute_force_knapsack_forfiling}}
\code{\link{brute_force_knapsack}}
\code{\link{knapsack_dynamic}}
}

