% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Solution for the knapsack problem by greedy}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{knapsack object as dataframe}

\item{W}{knapsack size}
}
\value{
result the list with maximum values and elements.
}
\description{
greedy_knapsack search, i.e. going through all possible alternatives and return the maximum value
found. This approach is of complexity O(2n) since all possible combinations 2n needs to be evaluated
}
