% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data frame with two columns:  v = value and w = weights of each item to put it into the knapsack.}

\item{W}{knapsack size, needs to be a positive value}
}
\value{
the maximum knapsack value and related elements
}
\description{
This algorithm proceeds to insert the items into the knapsack based on their value/weight ratio, inserting the items with higher ratios first. This method gives an approximation of the solution to the knapsack problem.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}
