% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyknapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{This is a description of greedy_knapsack function}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a data frame has two columns and all the elements are positive numbers}

\item{W}{a single number denotes the size of the knapsack}
}
\value{
a list contains the biggest value and the items
}
\description{
This function creates a greedy algorithm to solve the knapsack problem.
}
\examples{
#'set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n , replace = TRUE),
v = runif(n = n, 0 , 10000))
l <- greedy_knapsack(x = knapsack_objects[1:800, ], W = 3500)


}
