% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{knapsack_greedy

The greedy knapsack algorithm}
\arguments{
\item{x}{data frame with value and weights}

\item{W}{the capacity of the knapsack}
}
\value{
total value and elements
}
\description{
calculates the knapsack problem with the greedy algorithm.
}
\references{
https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm
}
