% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{is a data frame}

\item{W}{given max weight}
}
\value{
a list
}
\description{
knapsack problem being greedy check every possiblity of values.
}
\examples{
#' set.seed(42)
n <- 2000
knapsack_objects <-data.frame(
  w=sample(1:4000, size = n, replace = TRUE),
  v=runif(n = n, 0, 10000)
)
greedy_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
