% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack Solver}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with two columns: \code{w} for item weights and \code{v} for item values.}

\item{W}{An integer representing the maximum weight capacity of the knapsack.}
}
\value{
A list with two components:
\itemize{
\item \code{value}: The total value of the selected items.
\item \code{elements}: A vector of indices representing the selected items.
}
}
\description{
This function solves the knapsack problem using a greedy approach.
It selects items based on the value-to-weight ratio and attempts to fit as many items
as possible without exceeding the weight limit.
}
\examples{
knapsack_objects <- generate_knapsack_objects(10)
greedy_knapsack(knapsack_objects, W = 3500)

}
