% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic Algorithm for knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame with two elements, weights(w) and values(v)}

\item{W}{Knapsack capacity}
}
\value{
A list with optimal value and selected elements
}
\description{
Greedy Heuristic Algorithm for knapsack problem
}
