% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Data Frame that contain v(value) and w(weight)}

\item{W}{Vector eith the knapsack size}
}
\value{
a list with the value and elements of the knapsack
}
\description{
use the a heuristic or approximation for the problem
}
\examples{
set.seed(42) 
n <- 2000 
knapsack_objects <- data.frame(
w=sample(1:4000, size = n, replace = TRUE),
v=runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)

}
