% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Γreedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame that represents the wights and the values of each item.}

\item{W}{A number that represents the total weight .}
}
\value{
A list with the optimum value and the number of each item.
}
\description{
Γreedy_knapsack
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
