% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{dataframe with columns w and v (weight and value) representing the items}

\item{W}{double representing the weight limit of the knapsack}
}
\value{
list with the value of the items in the knapsack and the elements contained in the knapsack.
}
\description{
A greedy heuristic for solving the knapsack problem. 
The solution does not represent the optimal solution, but is computationally cheap.
}
