% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy function for knapsack problem}
\usage{
greedy_knapsack(x,W)
}
\arguments{
\item{x}{A data.frame. with weight and value}

\item{W}{A number. the overall weight}
}
\value{
list with maximum value and picked elements
}
\description{
implement the function using greedy method.
}
\examples{
  RNGversion(min(as.character(getRversion()),"3.6.1"))
  set.seed(42,kind="Mersenne-Twister",normal.kind = "Inversion")
  n <- 2000
  knapsack_objects <-data.frame(w=sample(1:4000, size = n, replace = TRUE),
                             v=runif(n = n, 0, 10000))
 l1 <- greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
