% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Title}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame containing values and weights of items}

\item{W}{The knapsack's capacity}
}
\value{
Maximum value and selected elements.
}
\description{
Title
}
