% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{An implementation of a greedy solution to the knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame with two vectors of the same size, v with values and w with weights}

\item{W}{an integer as the total weight of the knapsack}
}
\description{
An implementation of a greedy solution to the knapsack problem
}
