% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic solver for the Knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame with two columns named in order \code{"w"} and \code{"v"}
 Every row in \code{x} is an object with weight \code{w} and value \code{v}.}

\item{W}{A numeric scalar which is the limit of the weight the knapsack can carry.}
}
\value{
\code{list} A list with names \code{$value}, telling the maximum value of the knapsack
 and \code{$elements} which indicates which row objects in data.frame \code{x} was put
 in the knapsack.
}
\description{
This function solves the Knapsack problem using a greedy heuristic approach.
 It does not return an exact result but is computationally faster than the other
 approaches in this package.
}
\examples{
greedy_knapsack(data.frame(w=c(20,30,40,50),v=c(2,2,1,2)), W =200)
}
\references{
\url{http://en.wikipedia.org/wiki/Knapsack_problem}
}

