% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame cx with two variables v and w}

\item{W}{The weight capacity of the knapsack}
}
\value{
The maximum knapsack value and corresponding elements that contributes to the value.
}
\description{
Greedy knapsack implementation
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}
