% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6function.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{,a data.frame. x$w is the weight,x$v is the value}

\item{W}{,a integer. The maximum weight of items in the bag.}
}
\value{
list , returns a list containing the maximum value and the index of items in x.
}
\description{
greedy_knapsack for lab6,not very precise
}
