% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{the data set Knapsack_object}

\item{W}{the max weight allowed}
}
\description{
Using Greedy algorithm to solve
}
