% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_algorithm.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a dataframe}

\item{W}{the weight the knapsack can take,a interger}
}
\value{
a list contain the best value and specific plan(how to choose the items)
}
\description{
`greedy_knapsack()`solve the knapsack problem by greedy heuristic
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
  data.frame(
    w=sample(1:4000, size = n, replace = TRUE),
    v=runif(n = n, 0, 10000)
  )

greedy_knapsack(x = knapsack_objects, W = 10000)

}
\seealso{
https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm.
}
