% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Largest possible value in weighted bag - greedy}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data frame}

\item{W}{positive numeric value}
}
\value{
The output will be a numerical value(the largest possible sum) and the corresponding numerical elements that make that sum
}
\description{
This function takes the values and weights provided in a data frame and finds the largest possible sum of values that is less than or equal to the inputed weight by ordering each element by value per weight.
}
