% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{Solution for the knapsack problem}
\usage{
knapsack_brute_force(x, W)
}
\arguments{
\item{x}{snapsack object as dataframe}

\item{W}{snapsack size}
}
\value{
result the list with maximum values and elements.
}
\description{
knapsack_brute_force search, i.e. going through all possible alternatives and return the maximum value
found. This approach is of complexity O(2n) since all possible combinations 2n needs to be evaluated
}
