% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic Knapsack}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data frame with two columns, 'v' for values and 'w' for weights.}

\item{W}{The capacity of the knapsack.}
}
\value{
A list with the maximum value and a vector of selected item indices.
}
\description{
This function solves the 0/1 knapsack problem using dynamic programming.
}
