% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming solution to the knapsack problem.}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data frame with two variables, w and v.}

\item{W}{An integer which represent the weight limit for the knapsack problem.}
}
\value{
A list with the maximun value for a collection of items that fits under the weight
limit.
}
\description{
The knapsack_dynamic function applies a dynamic programing solution to the knapsack
problem. The function builds a matrix to store the calculations. 
The maximum value and the used itemsare returned by the function.
}
\examples{
##Get a solution for the knapsack problem on the data frame 
knapsack_objects, with the knapsack size of 2000.
knapsack_dynamic(x=knapsack_objects,W=2000)

}
\references{
More about the knapsack problem can be read here:
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}

