% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Brute-force search is guaranteed to give a correct answer in all situations for the knapsack problem.}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x, }{data.frame with with two variables weight(w) and value(v)}

\item{W, }{the size of knapsack}
}
\value{
the maximum knapsack value and which elements
}
\description{
This function calculates all possible alternatives and return the maximum value.
}
\details{
This function calculates in the easiest way to enumerate all different combinations.
}
\examples{
knapsack_dynamic(x = knapsack_objects[1:12,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 2000)
knapsack_dynamick(x = knapsack_objects[1:12,], W = 2000)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 2000)
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
\seealso{
\code{\link{brute_force_knapsack_forfiling}}
\code{\link{greedy_heuristic}}
\code{\link{brute_force_knapsack}}
}

