% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic knapsack}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data frame}

\item{W}{integer}
}
\value{
list
}
\description{
calulating possible weight capicity we can put in bag with help of discreat problem.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <-data.frame(
  w=sample(1:4000, size = n, replace = TRUE),
  v=runif(n = n, 0, 10000)
)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
}
