% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic Programming Knapsack Solver}
\arguments{
\item{x}{A data frame with two columns: \code{w} for item weights and \code{v} for item values.}

\item{W}{An integer representing the maximum weight capacity of the knapsack.}
}
\value{
A list with two components:
\itemize{
\item \code{value}: The maximum value of items that can be carried without exceeding the weight.
\item \code{elements}: A vector of indices representing the selected items.
}
}
\description{
This function solves the knapsack problem using dynamic programming.
It computes the optimal value and selected items for a given weight limit.
}
\examples{
knapsack_objects <- generate_knapsack_objects(500)
knapsack_dynamic(knapsack_objects, W = 3500)
}
