% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic approach to solve knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data frame with two elements, weights(w) and values(v)}

\item{W}{Knapsack capacity}
}
\value{
A list with optimal value and selected elements
}
\description{
Dynamic approach to solve knapsack problem
}
