% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming of knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A dataframe that contains weights and values.}

\item{W}{A numer representing the limited weight.}
}
\description{
A programme that solve the knapsack problem using dynamic programming.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
