% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{This is a function used to solve the knapsack problem with the dynamic 
programming solution}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{Data frame with all the elements that can be added to the knapsack.}

\item{W}{Weight limit of the knapsack.}
}
\value{
A list with the best computed value and the respective combination of
elements.
}
\description{
This function solves the knapsack problem using dynamic programming.
The inputs are a data frame of items with a value and a weight and a parameter
determining the total weight limit
}
\examples{
n = 2000
knapsack_objects <-
data.frame(
  w=sample(1:4000, size = n, replace = TRUE),
   v=runif(n = n, 0, 10000)
 )
knapsack_dynamic(knapsack_objects[1:12,],3500)
knapsack_dynamic(knapsack_objects[1:8,],2000)
}
