% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic_programming.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Implements Dynamic programming for knapsack problem}
\source{
https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem
}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame,with a weight w and a value v}

\item{W}{Knapsack capacity}
}
\value{
Maximum knapsack value and which elements (row in x)
}
\description{
solve the knapsack problem exact by iterating over all possible values of w
}
\details{
This function should return the same results as the brute force algorithm, but unlike the brute force it should scale much better since the algorithm will run in O(Wn).
}
