% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsac_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic algorithm for the Knapstack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data frame with two columns \code{w} (weight) and \code{v} (value) of items to place in the knapsack}

\item{W}{the maximum weight (numeric) the knapsack can hold}
}
\value{
theoretical maximum \code{$value} (knapsack value) composed of \code{$num} (which items)
}
\description{
Dynamic algorithm for the Knapstack problem
}
