% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming for knapsach probrem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data frame containing values and weights of items}

\item{W}{The knapsack's capacity}
}
\value{
Maximum value and selected elements.
}
\description{
Dynamic programming for knapsach probrem
}
