% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{An implementation of a dynamic solution to the knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame with two vectors of the same size, v with values and w with weights}

\item{W}{an integer as the total weight of the knapsack}
}
\description{
An implementation of a dynamic solution to the knapsack problem
}
