% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{the data set Knapsack_object}

\item{W}{the max weight allowed}
}
\description{
Using Dynamic programming algorithm to solve
}
