% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_programming.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic(x = knapsack_objects[1:500,], W = 3500)}
\usage{
knapsack_dynamic(x, W)
}
\description{
knapsack_dynamic(x = knapsack_objects[1:500,], W = 3500)
}
