% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{knapsack_greedy}
\alias{knapsack_greedy}
\title{Knapsack Greedy}
\usage{
knapsack_greedy(x, W, fast=FALSE)
}
\arguments{
\item{x}{A data frame containing the knapsack objects.
\describe{
\item{v}{The values of the objects.}
\item{w}{The weights of the objects.}
}}

\item{W}{The total amount of capacity available.}

\item{fast}{Set to FALSE by default. Will use a c++ implementation if set to TRUE.}
}
\value{
A list containing information about the solution.
\describe{
\item{value}{The best value found.}
\item{weight}{The weight of the best value.}
\item{elements}{The items picked to achieve the best solution.}
}
}
\description{
Solves 0-1 knapsack problem using greedy heuristic. The heuristic used is value / weight.
}
\examples{
knapsack_greedy(x = knapsack_objects(2000)[1:800,], W = 3500)
knapsack_greedy(x = knapsack_objects(2000)[1:800,], W = 3500, fast=TRUE)

}

