% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{This dataset is used for demonstrating the knapsack problem,}
\format{A data.frame with 2000 rows and two variables:
\describe{
 \item{w}{Contains weights}
 \item{v}{Contains values}
}}
\usage{
knapsack_objects
}
\description{
This dataset is used for demonstrating the knapsack problem,
}
\keyword{datasets}

