% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Example of an unweighted graph}
\format{
## `knapsack_objects`
A data frame with 2000 rows and 2 columns:
\describe{
  \item{w}{weight to put item into the knapsack}
  \item{v}{value of the item}
}
}
\usage{
knapsack_objects
}
\description{
A 'data.frame' with two variables (w and v) that contains the value (v) and the weight (w) of each item to put it into the knapsack.
}
\examples{
library(Lab6)
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
\keyword{datasets}
