% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Knapsack Objects}
\format{A data frame with two variables.
\describe{
\item{w}{The weights of the objects.}
\item{v}{The values of the objects.}
}}
\usage{
knapsack_objects(n)
}
\arguments{
\item{n}{The number of objects to create}
}
\value{
A data frame containing n knapsack objects
}
\description{
Function to generate knapsack objects
}

