% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Knapsack Objects}
\format{A data frame with two variables.
\describe{
\item{w}{The weights of the objects.}
\item{v}{The values of the objects.}
}}
\description{
A data set containing knapsack objects
}

