% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{knapsack_objects}
\format{
A dataframe with 2 columns and 2000 rows.
\describe{
\item{v}{Value of items}
\item{w}{weight of items}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\usage{
knapsack_objects
}
\description{
A dataset of a number of items' weights and values which is used in knapsack's problem.
}
\keyword{datasets}
