\name{lab6-package}
\alias{lab6-package}
\alias{lab6}
\docType{package}
\title{
\packageTitle{lab6}
}
\description{
\packageDescription{lab6}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab6}
\packageIndices{lab6}

}
\author{
\packageAuthor{lab6}

Maintainer: \packageMaintainer{lab6}
}
\references{

}

\keyword{ package }
\seealso{

\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{

}
