\name{lab6-package}
\docType{package}
\title{
\packageTitle{lab6}
}
\description{
\packageDescription{lab6}
}
\details{

The DESCRIPTION file
\packageDESCRIPTION{lab6}
\packageIndices{lab6}
create a package for the knapsack problem
}
\author{ packageAuthor {Jorge Villar Fuentes}

Maintainer packageMaintainer{Jorge Villar Fuentes}
}
