% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAB6.R
\docType{package}
\name{LAB6}
\alias{LAB6}
\alias{LAB6-package}
\title{The Knapsack Package}
\description{
With this package we can solve the knapsack problem with three ways, Greedy Heuristic, Dynamic Programming, Brute Force Search
}
\author{
Jorge Villar Fuentes and Ashmaful Alam
}
