\name{LAB6}
\docType{package}
\title{The Knapsack problem}
\packageTitle{LAB6}
}
\description{
\packageDescription{LAB6}
}
\details{

The DESCRIPTION file
\packageDESCRIPTION{LAB6}
\packageIndices{LAB6}
create a package of the knapsack problem
}
\author{ packageAuthor {Jorge Villar Fuentes}

Maintainer packageMaintainer{Jorge Villar Fuentes}
}
