\name{LAB6-package}
\alias{LAB6-package}
\alias{LAB6}
\docType{package}
\title{
\packageTitle{LAB6}
}
\description{
\packageDescription{LAB6}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LAB6}
\packageIndices{LAB6}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LAB6}

Maintainer: \packageMaintainer{LAB6}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
