% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6-package.R
\docType{package}
\name{lab6-package}
\alias{lab6-package}
\title{A package providing algorithms for solving the knapsack problem}
\description{
The package provides three different algorithms for solutions to the knapsack problem, 
each with different computational complexity. One provides the optimal solution, 
but is very computationally expensive, 
one is computationally cheaper but provides an optimal solution only for discrete weights 
and the third one is a heuristic which is computationally even cheaper, 
but the solution is not expected to be as good as for the other two algorithms.
}
\details{
lab6
}
