\name{Lab6-package}
\alias{Lab6-package}
\alias{Lab6}
\docType{package}
\title{
\packageTitle{Lab6}
}
\description{
\packageDescription{Lab6}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab6}
\packageIndices{Lab6}
}
\author{
\packageAuthor{Lab6}

Maintainer: \packageMaintainer{Lab6}
}
\keyword{ package }
