\name{Lab6-package}
\alias{Lab6-package}
\alias{Lab6}
\docType{package}
\title{
\packageTitle{Lab6}
}
\description{
\packageDescription{Lab6}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab6}
\packageIndices{Lab6}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lab6}

Maintainer: \packageMaintainer{Lab6}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
