% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6-package.R
\docType{package}
\name{Lab6-package}
\alias{Lab6}
\alias{Lab6-package}
\title{Lab6: Knapsack Problem}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
