% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\name{Lab6}
\alias{Lab6}
\title{Solve the Knapsack problem in no time}
\description{
This packages implements three different approaches to solving the knapsack problem. 
The first one is a brute force algorithm, a 0/1 dynamic solution and one greedy.
}

