% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab6.R
\docType{package}
\name{lab6}
\alias{lab6}
\alias{lab6-package}
\title{Knapsack problem solver!}
\description{
The package provides four categories of important functions:
}
\section{Lab6 functions}{

\code{\link{brute_force_knapsack}}
 
\code{\link{greedy_heuristic}} 

\code{\link{knapsack_dynamic}} 

\code{\link{brute_force_knapsack_profiling}}
}

\section{Dataset used in examples}{

\code{\link{knapsack_objects}}
}

