% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\docType{package}
\name{Lab6}
\alias{Lab6}
\alias{Lab6-package}
\title{Solving the knapsack problem.}
\description{
This package contains three different functions for solving what is called the knapsack problem.
}
\author{
Rasmus Holm

Zhendong Wang

Karolina Ziomek
}

