% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6.R
\docType{package}
\name{lab6}
\alias{lab6}
\title{Lab6}
\description{
Using different alogrithm to solve knapsack problem
}
\author{
Kyriakos Papadopoulos  \email{kyrpa853@student.liu.se} YiHung Chen  \email{yihch883@student.liu.se}
}
