% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomData.R
\name{RandomData}
\alias{RandomData}
\alias{randomData}
\title{RandomData}
\usage{
randomData()
}
\description{
This file aims to generate input data for knapsack
}
\examples{
knapsack_objects <- randomData()

}
