% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicProgramming.R
\name{updateStatus}
\alias{updateStatus}
\title{updateStatus}
\usage{
updateStatus(ssq, weights, values, w)
}
\arguments{
\item{ssq}{A numeric matrix to represents solutions of subproblems}

\item{weights}{A numeric variable represents weights .}

\item{values}{A numeric vector to values.}

\item{w}{A numeric variable represents the volume of the bag.}
}
\description{
This function aims to improve the performance in R implemented by C++.
}
